SETUP Add new directories and move corresponding files to them
