+++ src/components/App.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "./App.css";
+import "../styles/App.css";
 
 function App() {
   return (
