+++ src/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import App from "./App";
-import "./index.css";
+import App from "./components/App";
+import "./styles/index.css";
 
 ReactDOM.render(
   <React.StrictMode>
